+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.7.2] - 2020-05-dd
+
+**This release breaks backward compatibility with old versions! Existing databases need to be rebuilt.**
+
+### Changed
+
+- Seperated properties for referencing embedded and external artwork resources
+- Flattened content property in media source
+
 ## [0.7.1] - 2020-04-27
 
 **This release breaks backward compatibility with old versions! Existing databases need to be rebuilt.**